refactor: 모달의 동작 위치 변경
